User role wise access & update role

+++ app/Http/Controllers/AdminController.php
@@ -2,8 +2,10 @@
 
 namespace App\Http\Controllers;
 
+use App\Models\Role;
 use App\Models\User;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 
 class AdminController extends Controller
 {
@@ -18,6 +20,18 @@ class AdminController extends Controller
 
     }
 
+    public function authorize_admin()
+    {
+        $role = User::find(Auth::id())->role;
+        $user_level = $role->role_level;
+
+        if ($user_level == 3){
+            return true;
+        }
+
+        return 0;
+    }
+
     public function manageRoles()
     {
         return view('admin/roles');
@@ -25,17 +39,26 @@ class AdminController extends Controller
 
     public function manageUsers()
     {
-        $all_users = User::all();
+
+        if (!$this->authorize_admin()){
+            return abort(403);
+        }
+
+        $all_users = User::with('role')->get();
 
         return view('admin/users', ['users' => $all_users]);
     }
 
     public function deleteUser($id)
     {
+        if (!$this->authorize_admin()){
+            return abort(403);
+        }
+
         $deleted = User::destroy($id);
-        if ($deleted){
+        if ($deleted) {
             session()->flash('status', 'User Removed!');
-        }else{
+        } else {
             session()->flash('status', 'Task Failed!');
         }
 
@@ -44,25 +67,60 @@ class AdminController extends Controller
 
     public function updateUser(Request $request)
     {
+        if (!$this->authorize_admin()){
+            return abort(403);
+        }
+
         $user = $request->validate([
             'user_id' => 'required',
             'user_name' => 'required',
             'email' => 'required'
-        ]) ;
+        ]);
 
         $data_to_update['name'] = $user['user_name'];
         $data_to_update['email'] = $user['email'];
 
-        if (isset($request->password)){
+
+        if (isset($request->password)) {
             $data_to_update['password'] = bcrypt($request->password);
         }
 
+        // Update Or Create user role
+
+        if (($request->role) != -1) {
+            $role_name = ' ';
+
+            if ($request->role == 1) {
+                $role_name = "Seller";
+            } else if ($request->role == 2) {
+                $role_name = "Manager";
+            } else if ($request->role == 3) {
+                $role_name = "Admin";
+            } else if ($request->role == 0) {
+                $role_name = "Guest";
+            } else {
+// ---------------------make else rule
+                return 0;
+            }
+
+            $role_set = Role::updateOrCreate(
+                ['user_id' => $request->user_id],
+                ['role_name' => $role_name, 'role_level' => $request->role]
+            );
+
+            if ($role_set) {
+                session()->flash('status', 'Role Updated!');
+            }
+
+        }
+
+        // ------ update user details
         $status = User::where('id', $request->user_id)
             ->update($data_to_update);
 
-        if ($status){
+        if ($status) {
             session()->flash('status', 'User Updated!');
-        }else{
+        } else {
             session()->flash('status', 'Task Failed!');
         }
